reorg

+++ src/main/kotlin/no/nav/helseidnavtest/oppslag/adresse/AdresseRegisterClient.kt
@@ -11,6 +11,6 @@ class AdresseRegisterClient(private val adapter: AdresseRegisterCXFAdapter) {
     fun kommunikasjonsParter(fra: HerId, til: HerId) =
         Tjenester(kommunikasjonsPart(fra) as Tjeneste, kommunikasjonsPart(til) as Tjeneste)
 
-    fun kommunikasjonsPart(herId: HerId) = adapter.kommunikasjonsPart(herId.verdi)
+    fun kommunikasjonsPart(herId: HerId) = adapter.kommunikasjonsPart(herId.verdi.toInt())
 
 }
